Updated AFIs and instructions for FILTER2D_Lab

+++ module_02/filter2d/workspace/Filter2D/src/host/coefficients.h
@@ -1,6 +1,6 @@
 #pragma once
 
-const int filterCoeffs[][15][15] = {
+const short filterCoeffs[][15][15] = {
 	{
 		// filterCoeffs[0] -> Identity
 		{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0},
